Update userscript.user.js
It looks like all we have to do to make the censors work with the main site or the beta site is to remove these two lines.  Still haven't made the Neigh logo appear on the beta site yet, but that's gonna require some extra steps because that has effects instead of being static.